renamed generateAll to generateModules
