+++ src/genn/backends/cuda/optimiser.cc
@@ -29,7 +29,7 @@
 
 // GeNN code generator includes
 #include "code_generator/codeGenUtils.h"
-#include "code_generator/generateAll.h"
+#include "code_generator/generateModules.h"
 #include "code_generator/generateRunner.h"
 #include "code_generator/generateSynapseUpdate.h"
 #include "code_generator/generateSupportCode.h"
