+++ src/genn/generator/generator.cc
@@ -16,8 +16,8 @@
 #include "modelSpecInternal.h"
 
 // GeNN code generator includes
-#include "code_generator/generateAll.h"
 #include "code_generator/generateMakefile.h"
+#include "code_generator/generateModules.h"
 #include "code_generator/generateMSBuild.h"
 
 // Include backend
